Stop using old archive.py
This was a workaround for a python bug with zip files, but now we don't have any zip files being
downloaded and the bug is likely fixed. The new one causes weird
problems with the new salt, stop including it so that we use the updated
one shipped with salt instead.